requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
environ==1.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-environ==0.4.54
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.4.0
